+++ tests/apps/search/test_indexers_persons.py
@@ -96,6 +96,26 @@ class PersonsIndexersTestCase(TestCase):
             ],
         )
 
+    def test_indexers_persons_get_es_documents_language_fallback(self):
+        """Absolute urls should be computed as expected with language fallback."""
+        PersonFactory(
+            page_title={
+                "fr": "ma première personne",
+            },
+            should_publish=True,
+        )
+        indexed_persons = list(
+            PersonsIndexer.get_es_documents(index="some_index", action="some_action")
+        )
+
+        self.assertEqual(
+            indexed_persons[0]["absolute_url"],
+            {
+                "en": "/en/ma-premiere-personne/",
+                "fr": "/fr/ma-premiere-personne/",
+            },
+        )
+
     def test_indexers_persons_format_es_object_for_api(self):
         """
         Make sure format_es_object_for_api returns a properly formatted person