+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add support for language fallback on search indexes
 - Add support for language fallback to get page extensions directly and
   reversely related to a page via a plugin